release jira versions
Mark Jira versions as released once all other work is done. This is
slightly fraught as it's possible the GA build, which happens
asynchronously, can fail, but it feels like The Right Thing given that
the firm expectation that even if the current GA build fails, a
subsequent one will succeed.