ignore special chars ahead of Jira slugs
It sure would be nice if every commit had a title like
```
JIRA-123 Some nice comment
```

but lots of people like to wrap their Jira slugs in special characters
like `(JIRA-123)` or `[JIRA-123]` etc. Ignore those special characters.

Also, don't ignore the output when posting the release to Jira. If the
request fails, we want to know about it.