improve range-between-releases selection
Improve the way the "from" commit is selected: find the common ancestor
of the most recent tag and the head of the current branch. That is,
given a history that looks like
```
 -- v1 -- v1.1       -v2
/                   /
A -- B -- C -- D -- E
```
where the `v1.1` tag is not a direct ancestor to `v2`, it will correctly
identify `A` as the common ancestor, allowing A-E to be included when
searching for commits.